+++ src/components/sidebar/Sidebar.spec.jsx
@@ -11,7 +11,7 @@ import AppReducer from '@slices/AppSlice'
 import FriendReducer from '@slices/FriendSlice'
 import InviteReducer from '@slices/InviteSlice'
 import LobbyReducer from '@slices/LobbySlice'
-import MatchReducer from '@slices/MatchSlice'
+import MatchmakingReducer from '@slices/MatchmakingSlice'
 import NotificationReducer from '@slices/NotificationSlice'
 import UserReducer from '@slices/UserSlice'
 
@@ -106,7 +106,7 @@ describe('Sidebar Component', () => {
 
   const invites = []
 
-  const match = {
+  const matchmaking = {
     preMatch: null,
     match: null,
   }
@@ -134,7 +134,7 @@ describe('Sidebar Component', () => {
       user: UserReducer,
       notifications: NotificationReducer,
       invites: InviteReducer,
-      match: MatchReducer,
+      matchmaking: MatchmakingReducer,
       app: AppReducer,
       friends: FriendReducer,
       lobby: LobbyReducer,
@@ -145,7 +145,7 @@ describe('Sidebar Component', () => {
       notifications,
       invites,
       app,
-      match,
+      matchmaking,
       lobby,
     },
   })