+++ src/views/lobby/LobbyView.jsx
@@ -10,6 +10,7 @@ import {
   JoystickIcon,
   LobbyLineup,
   LobbyPlayButton,
+  MatchFoundModal,
 } from '@components'
 import { StorageService } from '@services'
 import { addToast } from '@slices/AppSlice'
@@ -20,12 +21,14 @@ import style from './LobbyView.module.css'
 export default function LobbyView() {
   const user = useSelector((state) => state.user)
   const lobby = useSelector((state) => state.lobby)
-  const preMatch = useSelector((state) => state.match.preMatch)
-  const match = useSelector((state) => state.match.match)
+  const preMatch = useSelector((state) => state.matchmaking.preMatch)
+
+  const match = useSelector((state) => state.matchmaking.match)
 
   const dispatch = useDispatch()
 
   const [secondsDiff, setSecondsDiff] = useState(null)
+  const [openMatchFoundModal, setOpenMatchFoundModal] = useState(false)
 
   const isOwner = lobby.owner_id === user.id
   const userPlayer = lobby.players?.find((player) => player.user_id === user.id)
@@ -34,7 +37,9 @@ export default function LobbyView() {
   )
 
   const handleQueue = async (action) => {
-    if (!isOwner || preMatch || match) return
+    if (preMatch || match) return
+
+    if (action === 'start' && !isOwner) return
 
     const userToken = StorageService.get('token')
 
@@ -77,7 +82,7 @@ export default function LobbyView() {
       }
     }
 
-    if (preMatch && preMatch.countdown === null) lockIn()
+    if (preMatch && preMatch.state === 'pre_start') lockIn()
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [preMatch])
 
@@ -106,7 +111,9 @@ export default function LobbyView() {
   }, [lobby])
 
   useEffect(() => {
-    if (lobby.queue) {
+    let intervalId
+
+    if (lobby.queue && !preMatch) {
       const date = DateTime.fromISO(lobby.queue.replace(' ', 'T'))
         .minus({ hours: 3 })
         .setZone('America/Sao_Paulo')
@@ -118,13 +125,20 @@ export default function LobbyView() {
         setSecondsDiff(diff)
       }
 
-      const interval = setInterval(calculateDiffInSeconds, 1000)
+      calculateDiffInSeconds()
 
-      return () => {
-        clearInterval(interval)
-      }
+      intervalId = setInterval(calculateDiffInSeconds, 1000)
     }
-  }, [lobby])
+
+    return () => {
+      clearInterval(intervalId)
+    }
+  }, [lobby, preMatch])
+
+  useEffect(() => {
+    if (preMatch && preMatch.state === 'lock_in') setOpenMatchFoundModal(true)
+    else setOpenMatchFoundModal(false)
+  }, [preMatch])
 
   return (
     <Container className={style.container} column gap={40} align="center">
@@ -141,9 +155,9 @@ export default function LobbyView() {
       </Container>
 
       {/*
-          TODO: move the following into its own component when
-          we have more then one game mode available
-        */}
+        TODO: move the following into its own component when
+        we have more then one game mode available
+      */}
       <Container className={style.gameType} fitContent>
         <Container
           className={style.gameTypeItem}
@@ -186,16 +200,24 @@ export default function LobbyView() {
           lobbyId={lobby.id}
         />
       </Container>
+
       <Container className={style.footer} fitContent>
         <LobbyPlayButton
           queueTime={lobby.queue && secondsDiff}
           restrictionCountdown={lobby.restriction_countdown}
           restricted={lobby.restriction_countdown}
+          disabled={!isOwner && !lobby.queue}
           onClick={
             lobby.queue_time !== null ? handleCancelQueue : handleStartQueue
           }
         />
       </Container>
+
+      <MatchFoundModal
+        isOpen={openMatchFoundModal}
+        setIsOpen={setOpenMatchFoundModal}
+        preMatch={preMatch}
+      />
     </Container>
   )
 }